pyproject.toml
[project]
name = "test-automation-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "allure-pytest>=2.13.5",
    "jenkins>=1.0.2",
    "pytest>=8.3.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "xlrd>=2.0.1",
    "xlutils>=2.0.0",
]
